moved Trello entities to another package
