+++ src/main/java/com/progress/tracking/rest/dto/trello/TrelloSearchResponse.java
@@ -1,0 +1,18 @@
+package com.progress.tracking.rest.dto.trello;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TrelloSearchResponse {
+
+    private List<Board> boards;
+
+}
