--- src/main/java/com/progress/tracking/wrapper/trello/pojo/Board.java
@@ -1,92 +1,0 @@
-package com.progress.tracking.wrapper.trello.pojo;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Board {
-
-    @SerializedName("id")
-    @Expose
-    private String id;
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-
-    @SerializedName("desc")
-    @Expose
-    private String desc;
-
-    @SerializedName("closed")
-    @Expose
-    private boolean closed;
-
-    @SerializedName("idOrganization")
-    @Expose
-    private String idOrganization;
-
-    @SerializedName("url")
-    @Expose
-    private String url;
-
-    @SerializedName("shortUrl")
-    @Expose
-    private String shortUrl;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDesc() {
-        return desc;
-    }
-
-    public void setDesc(String desc) {
-        this.desc = desc;
-    }
-
-    public boolean getClosed() {
-        return closed;
-    }
-
-    public void setClosed(boolean closed) {
-        this.closed = closed;
-    }
-
-    public String getIdOrganization() {
-        return idOrganization;
-    }
-
-    public void setIdOrganization(String idOrganization) {
-        this.idOrganization = idOrganization;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getShortUrl() {
-        return shortUrl;
-    }
-
-    public void setShortUrl(String shortUrl) {
-        this.shortUrl = shortUrl;
-    }
-
-}
