--- src/main/java/com/progress/tracking/wrapper/trello/pojo/Card.java
@@ -1,90 +1,0 @@
-package com.progress.tracking.wrapper.trello.pojo;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Card {
-    @SerializedName("id")
-    @Expose
-    private String id;
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-
-    @SerializedName("desc")
-    @Expose
-    private String description;
-
-    @SerializedName("idBoard")
-    @Expose
-    private String idBoard;
-
-    @SerializedName("idList")
-    @Expose
-    private String idList;
-
-    @SerializedName("url")
-    @Expose
-    private String url;
-
-    @SerializedName("shortUrl")
-    @Expose
-    private String shortUrl;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getIdBoard() {
-        return idBoard;
-    }
-
-    public void setIdBoard(String idBoard) {
-        this.idBoard = idBoard;
-    }
-
-    public String getIdList() {
-        return idList;
-    }
-
-    public void setIdList(String idList) {
-        this.idList = idList;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getShortUrl() {
-        return shortUrl;
-    }
-
-    public void setShortUrl(String shortUrl) {
-        this.shortUrl = shortUrl;
-    }
-}
