--- src/main/java/com/progress/tracking/wrapper/trello/pojo/TrelloList.java
@@ -1,52 +1,0 @@
-package com.progress.tracking.wrapper.trello.pojo;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class TrelloList {
-
-    @SerializedName("id")
-    @Expose
-    private String id;
-    @SerializedName("name")
-    @Expose
-    private String name;
-    @SerializedName("closed")
-    @Expose
-    private boolean closed;
-    @SerializedName("idBoard")
-    @Expose
-    private String idBoard;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public boolean getClosed() {
-        return closed;
-    }
-
-    public void setClosed(boolean closed) {
-        this.closed = closed;
-    }
-
-    public String getIdBoard() {
-        return idBoard;
-    }
-
-    public void setIdBoard(String idBoard) {
-        this.idBoard = idBoard;
-    }
-}
